Make pcb-sch Starlark bindings optional
Amp-Thread-ID: https://ampcode.com/threads/T-01a084a7-17f2-7198-abf7-b6979dfa0e9a
